Turns out that having a save button is required for forms ;)
For a save button to work with the Options API, the option needs to be
registered globally, hence the new admin-global class.

It's also a lot easier to find the New SEO menu when it's in the same
spot all the time, so registering that in the global admin class too.
To prevent high memory usage for stuff that isn't used, it uses a dummy
function and only in admin.php type pages is the real class being
loaded.